Moved location code into it's own class. Updated permission helper
